Move features folder to server-interface
