+++ src/protocol/README.md
@@ -2,10 +2,3 @@
 
 This folder contains definition of AWS Language Servers Protocol, which defines a protocol for communication between Servers and Clients.
 It is modelled after LSP, and provides several custom extensions to enable custom AWS Language Server Runtimes features.
-
-
-```
-
-Language Server Capability <===  (Capability Features interface) ===> Language Server Runtime <===(Runtimes Protocol (LSP extensions))===> Language Client
-
-```
