+++ src/protocol/lsp.ts
@@ -1,4 +1,10 @@
 export { TextDocument, Position } from 'vscode-languageserver-textdocument'
+
+// LSP protocol is a core dependency for LSP feature provided by runtimes.
+// Since we aim to provide whole range of LSP specification for Clients and Capabilities,
+// we re-exporting whole LSP protocol for usage.
+// Scoping it down is not practical due to large surface of protocol and types relationship.
+// It will be limiting implementors, if they choose to type their code with more specific types from LSP.
 export * from 'vscode-languageserver-protocol'
 
 // Custom Runtimes LSP extensions
