+++ src/runtimes/index.ts
@@ -1,2 +1,1 @@
-export { Server } from '../features/server'
 export { standalone } from './standalone'
