+++ src/runtimes/lsp/index.ts
@@ -1,0 +1,1 @@
+export { observe } from './textDocuments/textDocumentConnection'
