+++ src/server-interface/README.md
@@ -1,0 +1,3 @@
+## AWS Language Server types
+
+This module contains types, that define interfaces of features available to Server implementors and used by AWS Runtimes to provide it's features.
