+++ src/server-interface/index.ts
@@ -1,6 +1,6 @@
 export { Server } from './server'
-export * from './auth/auth'
+export * from './auth'
 export { Logging } from './logging'
 export * from './lsp'
-export * from './telemetry/telemetry'
+export * from './telemetry'
 export { Workspace } from './workspace'
