+++ src/server-interface/lsp.ts
@@ -21,13 +21,11 @@ import {
     InlineCompletionListWithReferences,
     LogInlineCompletionSessionResultsParams,
     RequestHandler,
-} from '../../protocol'
+} from '../protocol'
 
 // Re-export whole surface of LSP protocol used in Runtimes.
 // This is needed for LSP features as we pass messages down.
-export * from '../../protocol/lsp'
-
-export { observe } from './textDocuments/textDocumentConnection'
+export * from '../protocol/lsp'
 
 export type PartialServerCapabilities<T = any> = Pick<ServerCapabilities<T>, 'completionProvider' | 'hoverProvider'>
 export type PartialInitializeResult<T = any> = {
